define server GET / api
