--- server.js
@@ -1,9 +1,0 @@
-const express = require('express');
-const helmet = require('helmet');
-
-const server = express();
-
-server.use(express.json());
-server.use(helmet());
-
-module.exports = server;
